fix(skills-service): fix employees having duplicate rankings
Adds changes to the getEmployeeRankings() API which now continuously ranks employees
and removes duplicate rankings if the scores are similar. First employee to reach
that ranking will now stay on that rank and any new employees to reach that level
will stay below that rank unless they achieve a score higher than the employee
above them.